# Support for Native Quadforms (x.T @ A @ x) 