--- a/cvxpy/atoms/quad_form.py
+++ b/cvxpy/atoms/quad_form.py
@@ -54,7 +54,7 @@ class QuadForm(Atom):
         if self.args[1].shape[1] != n or self.args[0].shape not in [(n, 1), (n,)]:
             raise ValueError("Invalid dimensions for arguments.")
         if not self.args[1].is_hermitian():
-            raise ValueError("P must be symmetric/Hermitian.")
+            raise ValueError("Quadratic form matrices must be symmetric/Hermitian.")
 
     def sign_from_args(self) -> Tuple[bool, bool]:
         """Returns sign (is positive, is negative) of the expression.