Hi -- small QOL upgrade for my personal cvxpy usage to automatically turn native matmul into quadratic forms where applicable. See issue #1762. When expressions are being constructed, then some MulExpression are automatically turned into a QuadForm, when certain criteria hold. 


Notes: 
1) Does not handle associativity. If we have a constant vector c, then `1/2 * c.T @ c @ x.T @ A @ x` is technically a quadratic form `(1/2 * c.T @ c) * QuadForm(x, A)`, but will not be converted in the below change. 
Reasoning here is that the expression tree leading up to the quadratic form can be arbitrarily tall and would require a search for the right-most leaf (to avoid QuadForms in cases like `1/2 * (c.T @ (c @ (d.T @ (d @ x.T)))) @ A @ y`).  However, as long as parantheses are included around `x.T @ A @ x`, the Expr will have a correct QuadForm leaf.  

2) Does not handle asymmetric/non-hermitian matrices. I noticed that the default QuadForm behavior currently is to throw an error if QuadForm(x, A) is called with A being asymmetric/nonhermitian. Could simply replace A with (A + A.H) / 2 but saw that was removed a while back. 

Let me know if these are not the desired behavior -- thanks! 


Best,
Jeffrey


## Type of change
- [x] New feature (backwards compatible)
- [ ] New feature (breaking API changes)
- [ ] Bug fix
- [ ] Other (Documentation, CI, ...)

## [Contribution checklist](https://www.cvxpy.org/contributing/index.html#contribution-checklist)
- [ ] Add our license to new files. (N/A) 
- [x] Check that your code adheres to our coding style.
- [x] Write unittests.
- [x] Run the unittests and check that they’re passing.
- [x] Run the benchmarks to make sure your change doesn’t introduce a regression.

<details>
  <summary>Benchmark Results</summary>

Without Changes: 

```
============================================================================================================================== test session starts ==============================================================================================================================
platform darwin -- Python 3.9.7, pytest-6.2.5, py-1.11.0, pluggy-1.0.0
rootdir: /Users/jchen/Desktop/cvxpy, configfile: pyproject.toml
plugins: dash-2.1.0, hypothesis-6.24.1, cov-3.0.0
collected 12 items                                                                                                                                                                                                                                                              

cvxpy/tests/test_benchmarks.py cone_matrix_stuffing_with_many_constraints: avg=1.159e+00 s , std=0.000e+00 s (1 iterations)
.diffcp_sdp: avg=6.678e+00 s , std=0.000e+00 s (1 iterations)
.-------------------------------------------------------------------------------
                                  Compilation                                  
-------------------------------------------------------------------------------
(CVXPY) Dec 25 01:40:42 PM: Compiling problem (target solver=ECOS).
(CVXPY) Dec 25 01:40:42 PM: Reduction chain: Dcp2Cone -> CvxAttr2Constr -> ConeMatrixStuffing -> ECOS
(CVXPY) Dec 25 01:40:42 PM: Applying reduction Dcp2Cone
(CVXPY) Dec 25 01:40:42 PM: Applying reduction CvxAttr2Constr
(CVXPY) Dec 25 01:40:42 PM: Applying reduction ConeMatrixStuffing
(CVXPY) Dec 25 01:40:43 PM: Applying reduction ECOS
(CVXPY) Dec 25 01:40:43 PM: Finished problem compilation (took 6.253e-01 seconds).
Issue #1668 regression test
Compilation time:  0.6262228488922119
.least_squares: avg=4.767e-03 s , std=0.000e+00 s (1 iterations)
.sConic canonicalization
(ECOS) solver time:  0.059706222
cvxpy time:  0.8578147779084473
Quadratic canonicalization
(OSQP) solver time:  0.02199917
cvxpy time:  0.932844112699585
.qp: avg=8.303e-03 s , std=0.000e+00 s (1 iterations)
.small_cone_matrix_stuffing: avg=4.381e-02 s , std=8.711e-03 s (10 iterations)
.small_lp: avg=1.545e-02 s , std=0.000e+00 s (1 iterations)
small_lp_second_time: avg=1.258e-03 s , std=0.000e+00 s (1 iterations)
.sstv_inpainting: avg=4.951e+00 s , std=0.000e+00 s (1 iterations)
.

=============================================================================================================================== warnings summary ================================================================================================================================
cvxpy/tests/test_benchmarks.py::TestBenchmarks::test_parameterized_qp
  /Users/jchen/Desktop/cvxpy/cvxpy/reductions/solvers/solving_chain.py:209: UserWarning: Your problem has too many parameters for efficient DPP compilation. We suggest setting 'ignore_dpp = True'.
    warnings.warn(

-- Docs: https://docs.pytest.org/en/stable/warnings.html
```
  
With Changes:
```

============================================================================================================================== test session starts ==============================================================================================================================
platform darwin -- Python 3.9.7, pytest-6.2.5, py-1.11.0, pluggy-1.0.0
rootdir: /Users/jchen/Desktop/cvxpy, configfile: pyproject.toml
plugins: dash-2.1.0, hypothesis-6.24.1, cov-3.0.0
collected 12 items                                                                                                                                                                                                                                                              

cvxpy/tests/test_benchmarks.py cone_matrix_stuffing_with_many_constraints: avg=1.166e+00 s , std=0.000e+00 s (1 iterations)
.diffcp_sdp: avg=6.146e+00 s , std=0.000e+00 s (1 iterations)
.-------------------------------------------------------------------------------
                                  Compilation                                  
-------------------------------------------------------------------------------
(CVXPY) Dec 25 01:40:15 PM: Compiling problem (target solver=ECOS).
(CVXPY) Dec 25 01:40:15 PM: Reduction chain: Dcp2Cone -> CvxAttr2Constr -> ConeMatrixStuffing -> ECOS
(CVXPY) Dec 25 01:40:15 PM: Applying reduction Dcp2Cone
(CVXPY) Dec 25 01:40:15 PM: Applying reduction CvxAttr2Constr
(CVXPY) Dec 25 01:40:15 PM: Applying reduction ConeMatrixStuffing
(CVXPY) Dec 25 01:40:15 PM: Applying reduction ECOS
(CVXPY) Dec 25 01:40:16 PM: Finished problem compilation (took 5.758e-01 seconds).
Issue #1668 regression test
Compilation time:  0.5766229629516602
.least_squares: avg=4.683e-03 s , std=0.000e+00 s (1 iterations)
.sConic canonicalization
(ECOS) solver time:  0.0435112
cvxpy time:  0.8134698390472412
Quadratic canonicalization
(OSQP) solver time:  0.019287906
cvxpy time:  0.873811163595337
.qp: avg=7.598e-03 s , std=0.000e+00 s (1 iterations)
.small_cone_matrix_stuffing: avg=3.792e-02 s , std=8.338e-03 s (10 iterations)
.small_lp: avg=1.440e-02 s , std=0.000e+00 s (1 iterations)
small_lp_second_time: avg=1.113e-03 s , std=0.000e+00 s (1 iterations)
.sstv_inpainting: avg=3.103e+00 s , std=0.000e+00 s (1 iterations)
.

=============================================================================================================================== warnings summary ================================================================================================================================
cvxpy/tests/test_benchmarks.py::TestBenchmarks::test_parameterized_qp
  /Users/jchen/Desktop/cvxpy/cvxpy/reductions/solvers/solving_chain.py:209: UserWarning: Your problem has too many parameters for efficient DPP compilation. We suggest setting 'ignore_dpp = True'.
    warnings.warn(

-- Docs: https://docs.pytest.org/en/stable/warnings.html
```

</details>