@SteveDiamond 2022-12-26
This is a great contribution and well implemented! We are discussing internally what sort of additional error messages might be appropriate related to this feature. We will get back to you.