@SteveDiamond 2022-12-27
I think our biggest question is whether the different treatment of symmetric and non-symmetric matrices would be confusing. It might be better to map quadratic forms to quad_form regardless of whether the matrix is symmetric or not, and modify the error message for non-symmetric inputs to quad_form. @PTNobel what do you think?