@SteveDiamond 2022-12-27
> Very excited about this feature!
> 
> First question, is there any situation in CVXPY where we would expect `x.T @ A @ x` with a non-symmetric `A` to not raise an error? If so, I think we should just raise an error about the symmetry instead of raising a DCP error later.

In the long run we might want to allow ``x.T @ A @ x`` with non-symmetric ``A`` and symmetrize ``A``. But for now I agree it should raise an error right away. Currently in this PR ``x.T @ A @ x`` doesn't raise an error but is non-DCP. ``quad_form(x, A)`` raises an error immediately for non-symmetric ``A``.

> 
> Second question, which I can answer myself tonight when I get back to my computer what do the following errors look like:
> 
> * `x.T @ A @ x` for asymmetric `A`

DCP error
> * `x.T @ A @ x` for symmetric but not PSD `A`

DCP error