@rileyjmurray 2022-12-27
Hi @jlchen0, thanks so much for this contribution! As food for thought when it comes to error messages, here is some undesirable behavior that CVXPY currently implements:

```
import numpy as np
import cvxpy as cp

C = np.array([[1,0],[1,1]])
X = cp.Variable(shape=(2,2))
objective = cp.Minimize(cp.norm(C - X))

cp.Problem(objective, [X >> 0]).solve()
print(X.value.round(decimals=3))
# array([[ 1., -0.],
#        [ 1.,  1.]])

Problem(objective, [cp.lambda_min(X) >= 0]).solve()
print(X.value.round(decimals=3))
# array([[1. , 0.5],
#        [0.5, 1. ]])
```

This example shows that assuming symmetry (or working with the symmetric part of an input) results in inconsistent behavior with ``X >> 0`` and ``lambda_min(X) >= 0``.

So my preference would be to enforce symmetry of A, as with the current implementation of ``quad_form``.