@rileyjmurray 2022-12-27
I suggest that we use a ``ValueError`` for ``x.T @ A @ x`` when ``A`` is symmetric, with a warning that quadratic forms should only be defined with symmetric matrices. I think a DCP error should be raised later on (during compilation) if ``A`` is indefinite. My reasoning here is that someone might want to define a library that extends CVXPY to support nonconvex quadratic forms, which solvers like GUROBI already support. We shouldn't get in the way of that kind of valid extension if someone wants to implement it outside of CVXPY.