@jlchen0 2022-12-27
Thanks everyone for the feedback! Most recent commit now causes `x.T @ A @ x` to throw a `ValueError` if all conditions are met except symmetry/hermitian-ness. This is done by constructing a QuadForm regardless of symmetry, and then deferring to the `ValueError` thrown by QuadForm for asymmetric matrices. 

Also updated the error message to be a bit more clear and an associated test for it. 

Thanks!